move inbound index metrics into module
